converted routes to a module
