+++ Commander/tests/server_tests/test_agentCheckin.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import json
 from server import jobsCache
-from server.routes import agentCheckin
+from server.routes.agent import agentCheckin
 from utils import timestampToDatetime
